Refactor DestinationController to use DestinationService with injected RapidAPI config and HttpClient
