+++ src/GoTorz.Api/Services/IDestinationService.cs
@@ -1,0 +1,10 @@
+﻿using GoTorz.Shared.DTOs;
+
+namespace GoTorz.Api.Services
+{
+    public interface IDestinationService
+    {
+        Task<List<DestinationDto>> SearchDestinationAsync(string query);
+    }
+}
+
